Fully integrated Graphql api
